+++ src/bootstrap/index.ts
@@ -1,3 +1,3 @@
 export { bootstrap } from './functions/bootstrap.function';
 export type { IAppConfig, IMelodicApp } from './interfaces';
-export * from './types';
+export type * from './types';