+++ src/config/index.ts
@@ -1,4 +1,9 @@
-export * from './environment';
-export * from './define-config';
-export * from './injection-tokens';
-export * from './provide-config';
+export { environment, getEnvironment } from './environment';
+export { defineConfig } from './define-config';
+export { APP_CONFIG } from './injection-tokens';
+export { provideConfig } from './provide-config';
+
+export type * from './environment';
+export type * from './define-config';
+export type * from './injection-tokens';
+export type * from './provide-config';