+++ src/http/index.ts
@@ -1,4 +1,7 @@
-export * from './classes';
-export * from './functions';
+export { AbortError, HttpBaseError, HttpClient, HttpError, NetworkError } from './classes';
+export { provideHttp } from './functions';
+
+export type * from './classes';
+export type * from './functions';
 export type * from './interfaces';
 export type * from './types';