+++ src/index.ts
@@ -1,11 +1,120 @@
-export * from './bootstrap';
-export * from './components';
-export * from './config';
-export * from './forms';
-export * from './http';
-export * from './injection';
-export * from './interfaces';
-export * from './routing';
-export * from './signals';
-export * from './state';
-export * from './template';
+// Bootstrap
+export { bootstrap } from './bootstrap';
+export type * from './bootstrap';
+
+// Components
+export { ComponentBase, MelodicComponent, applyGlobalStyles, getActiveComponent, setActiveComponent } from './components';
+export type * from './components';
+
+// Config
+export { APP_CONFIG, defineConfig, environment, getEnvironment, provideConfig } from './config';
+export type * from './config';
+
+// Forms
+export {
+	AbstractControl,
+	FormArray,
+	FormControl,
+	FormGroup,
+	Validators,
+	checkboxAdapter,
+	createAsyncValidator,
+	createFormArray,
+	createFormControl,
+	createFormGroup,
+	createValidator,
+	formControlDirective,
+	getAdapter,
+	getGlobalMessage,
+	radioAdapter,
+	registerAdapter,
+	registerDefaultMessages,
+	resolveMessage,
+	setDefaultMessage,
+	textAdapter
+} from './forms';
+export type * from './forms';
+
+// HTTP
+export { AbortError, HttpBaseError, HttpClient, HttpError, NetworkError, provideHttp } from './http';
+export type * from './http';
+
+// Injection
+export { Binding, Inject, Injectable, InjectionEngine, Injector, Service, createToken, describeToken, getTokenKey } from './injection';
+export type * from './injection';
+
+// Interfaces
+export type * from './interfaces';
+
+// Routing
+export {
+	ROUTE_CONTEXT_EVENT,
+	RouteContextEvent,
+	RouteContextService,
+	RouteMatcher,
+	RouterLinkComponent,
+	RouterLinkCore,
+	RouterOutletComponent,
+	RouterService,
+	buildPathFromRoute,
+	createDeactivateGuard,
+	createGuard,
+	createResolver,
+	findRouteByName,
+	installHistoryEvents,
+	isSafeUrl,
+	matchRouteTree,
+	provideRouter,
+	routerLinkDirective
+} from './routing';
+export type * from './routing';
+
+// Signals
+export { SIGNAL_MARKER, SignalEffect, batch, computed, getActiveEffect, isSignal, setActiveEffect, signal } from './signals';
+export type * from './signals';
+
+// State
+export {
+	ComponentStateBaseService,
+	EffectsBase,
+	RX_ACTION_PROVIDERS,
+	RX_EFFECTS_PROVIDERS,
+	RX_INIT_STATE,
+	RX_STATE_DEBUG,
+	SignalStoreService,
+	createAction,
+	createReducer,
+	createState,
+	onAction,
+	props,
+	provideRX
+} from './state';
+export type * from './state';
+
+// Template
+export {
+	Directive,
+	TemplateResult,
+	classMap,
+	css,
+	directive,
+	disposeContainerParts,
+	disposeDirectiveState,
+	disposePart,
+	disposeParts,
+	getAttributeDirective,
+	getRegisteredDirectives,
+	hasAttributeDirective,
+	html,
+	isDirective,
+	portalDirective,
+	registerAttributeDirective,
+	render,
+	repeat,
+	repeatRaw,
+	styleMap,
+	unregisterAttributeDirective,
+	unsafeHTML,
+	when
+} from './template';
+export type * from './template';