+++ src/injection/index.ts
@@ -1,6 +1,11 @@
-export * from './decorators';
-export * from './function';
-export * from './interfaces';
-export * from './types';
-export * from './classes/binding.class';
-export * from './classes/injection-engine.class';
+export { Inject, Injectable, Service } from './decorators';
+export { createToken, describeToken, getTokenKey } from './function';
+export { Binding } from './classes/binding.class';
+export { InjectionEngine, Injector } from './classes/injection-engine.class';
+
+export type * from './decorators';
+export type * from './function';
+export type * from './interfaces';
+export type * from './types';
+export type * from './classes/binding.class';
+export type * from './classes/injection-engine.class';