+++ src/signals/index.ts
@@ -1,3 +1,7 @@
-export * from './classes/signal-effect.class';
-export * from './functions';
-export * from './types';
+export { SignalEffect } from './classes/signal-effect.class';
+export { batch, computed, getActiveEffect, isSignal, setActiveEffect, signal } from './functions';
+export { SIGNAL_MARKER } from './types';
+
+export type * from './classes/signal-effect.class';
+export type * from './functions';
+export type * from './types';